+++ getting_started/test_renderer.py
@@ -1,8 +1,11 @@
 import grid2op
-from grid2op.Agent import DoNothingAgent
-
+from grid2op.Agent import DoNothingAgent, RandomAgent, MLAgent
+from grid2op.Converters import IdToAct
+import pdb
 env = grid2op.make("case5_example")
-my_agent = DoNothingAgent(env.action_space)
+# my_agent = DoNothingAgent(env.action_space)
+# my_agent = RandomAgent(env.action_space)
+my_agent = MLAgent(env.action_space)
 
 all_obs = []
 obs = env.reset()
@@ -16,7 +19,9 @@ while True:
     action = my_agent.act(obs, reward, done)
     obs, reward, done, _ = env.step(action)
     print("Rendering timestep {}".format(nb_step))
-    env.render()
+    arr_ = env.render()
+#     arr_ = env.render(mode="rgb_array")
+#     print(arr_.shape)
     if done:
         break
     all_obs.append(obs)