+++ getting_started/test_renderer_14.py
@@ -1,6 +1,6 @@
 import grid2op
 from grid2op.Agent import DoNothingAgent
-from grid2op.Agent import GreedyAgent
+from grid2op.Agent import GreedyAgent, RandomAgent
 import numpy as np
 
 env = grid2op.make("l2rpn_2019", chronics_path="/home/donnotben/Téléchargements/L2RPN_pypownetBundle/G2Op_L2RPN/starting_kit_1/public_data/L2RPN2019_chronics")
@@ -100,7 +100,9 @@ class MyExpertAgent(GreedyAgent):
         return res
     
 my_agent = MyExpertAgent(env.action_space)
-
+my_agent = RandomAgent(env.action_space)
+print("Total unitary action possible: {}".format(my_agent.action_space.n))
+      
 all_obs = []
 obs = env.reset()
 all_obs.append(obs)